requirements.txt
absl-py==0.7.1
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
astor==0.7.1
astunparse==1.6.3
async-timeout==3.0.1
atari-py==0.1.15
atomicwrites==1.3.0
attrs==19.3.0
audioread==2.1.6
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
bleach==1.5.0
blessings==1.7
cachetools==4.1.1
cattrs==1.0.0
certifi==2019.3.9
chardet==3.0.4
chumpy==0.70
click==8.0.1
cloudpickle==0.8.1
colorama==0.4.4
contextvars==2.4
cycler==0.10.0
Cython==0.29.22
dataclasses==0.8
decorator==4.4.0
deepdish==0.3.6
defusedxml==0.6.0
dill==0.2.9
Django==2.2.1
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
flatbuffers==1.12
fsspec==0.8.4
future==0.17.1
gast==0.3.3
glob2==0.6
google-api-core==1.28.0
google-auth==1.30.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
grpcio==1.34.0
gym==0.12.5
h5py==2.10.0
hiredis==2.0.0
html5lib==0.9999999
idna==2.8
idna-ssl==1.1.0
image==1.5.25
imageio==2.5.0
immutables==0.15
importlib-metadata==4.3.0
imutils==0.5.2
ipdb==0.12
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==2.0.1
jedi==0.13.3
Jinja2==3.0.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kaleido==0.2.1
Keras-Preprocessing==1.1.2
kiwisolver==1.0.1
librosa==0.6.3
llvmlite==0.28.0
Markdown==3.1
MarkupSafe==2.0.1
matplotlib==2.2.2
mistune==0.8.4
mock==3.0.5
more-itertools==7.0.0
mpi4py==3.0.3
msgpack==1.0.2
multidict==5.1.0
munch==2.5.0
munkres==1.0.12
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.3
nibabel==2.4.0
notebook==5.7.8
numba==0.43.1
numexpr==2.6.9
numpy==1.19.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus==0.7.13
opencensus-context==0.1.2
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.1.2
pkg-resources==0.0.0
plotly==4.14.3
pluggy==0.11.0
prometheus-client==0.10.1
prompt-toolkit==2.0.9
protobuf==3.17.1
psutil==5.8.0
ptyprocess==0.6.0
py==1.8.0
py-spy==0.3.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.3.2
Pygments==2.3.1
PyOpenGL==3.1.0
pyparsing==2.3.1
pyrsistent==0.14.11
pytest==3.10.1
python-dateutil==2.8.0
pytorch-lightning==1.0.8
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
ray==1.3.0
redis==3.5.3
reprint==0.5.2
requests==2.21.0
requests-oauthlib==1.3.0
resampy==0.2.1
retrying==1.3.3
rsa==4.6
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.16.0
sklearn==0.0
sqlparse==0.3.0
stable-baselines==2.5.1
tables==3.5.1
tabulate==0.8.9
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorboardX==2.2
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.7.0
torchfile==0.1.0
torchvision==0.8.1
tornado==6.0.2
tqdm==4.54.1
traitlets==4.3.2
typing-extensions==3.7.4.3
urllib3==1.24.1
visdom==0.1.8.8
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==2.0.1
widgetsnbextension==3.4.2
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
zmq==0.0.0